- krijimi i faqes workspaces dhe ndryshime ne disa metoda
- krijimi i faqes workspaces ku shfaqen workspaces te nje useri dhe klasifikohen ato ku ai eshte owner dhe ku eshte member i thjeshte si dhe ka butonat per te naviguar te boards, members dhe settings
- gjithashtu eshte ndryshuar metoda getAll qe te kthej dto te members si dhe eshte ndrryshuar metoda delete board ashtu qe kur te fshihet nje bord te fshihet edhe nga tabela e starred

+++ backend/Repositories/BoardRepository.cs
@@ -14,10 +14,12 @@ namespace backend.Repositories
     public class BoardRepository : IBoardRepository
     {
         private readonly ApplicationDBContext _context;
+  
         
         public BoardRepository(ApplicationDBContext context)
         {
             _context=context;
+          
         }
         
         //GetAllAsync --> Gets all boards that exists
@@ -77,7 +79,12 @@ namespace backend.Repositories
 
             if (boardModel == null)
                 return null;
-
+            var starredBoard = await _context.StarredBoard.FirstOrDefaultAsync(x => x.BoardId == id);
+            
+            if (starredBoard == null)
+                return null;
+            
+            _context.StarredBoard.Remove(starredBoard);
             _context.Board.Remove(boardModel);
             await _context.SaveChangesAsync();
             return boardModel;
